requirements.txt
pymupdf
customtkinter
pillow
pytesseract
